feat: add reconcile-live CLI command for orphaned Alpaca position detection
Adds a reconciliation tool that compares live Alpaca positions against the
shadow_trades database, backfilling orphaned positions and marking stale
records. Fixes fractional share truncation in alpaca_adapter (int → float).

- New `reconcile-live` CLI command with --dry-run flag
- New `src/shadow_trading/reconcile.py` module
- POST /api/live/reconcile endpoint (cloud returns CLI instruction)
- LiveLedger.jsx: disabled Reconcile button with tooltip
- alpaca_adapter: float(qty) in all 3 position functions (fractional shares)
- 5 new tests (dry-run, backfill, stale, no-discrepancy, paper isolation)

+++ src/shadow_trading/alpaca_adapter.py
@@ -236,7 +236,7 @@ def get_position(ticker: str) -> dict | None:
         pos = client.get_open_position(ticker)
         return {
             "symbol": str(pos.symbol),
-            "qty": int(pos.qty),
+            "qty": float(pos.qty),
             "avg_entry_price": float(pos.avg_entry_price),
             "current_price": float(pos.current_price),
             "market_value": float(pos.market_value),
@@ -254,7 +254,7 @@ def get_all_positions() -> list[dict]:
     return [
         {
             "symbol": str(pos.symbol),
-            "qty": int(pos.qty),
+            "qty": float(pos.qty),
             "avg_entry_price": float(pos.avg_entry_price),
             "current_price": float(pos.current_price),
             "market_value": float(pos.market_value),
@@ -468,7 +468,7 @@ def get_live_positions() -> list[dict]:
     return [
         {
             "symbol": str(pos.symbol),
-            "qty": int(pos.qty),
+            "qty": float(pos.qty),
             "avg_entry_price": float(pos.avg_entry_price),
             "current_price": float(pos.current_price),
             "market_value": float(pos.market_value),
